requirements.txt
asgiref==3.9.1
channels==4.2.2
channels_redis==4.2.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
joblib==1.5.1
Khayyam==3.0.17
msgpack==1.1.1
numpy==2.3.1
pandas==2.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.2.0
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
